+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - added --samples command line option to spiders.py
+- ```cloudfeaster.webdriver_spider.WebElement``` now has
+a ```is_element_present()``` method that functions just
+like ```cloudfeaster.webdriver_spider.Browser```
 
 ### Changed